Correct fix for blocking DriverAccepts when acceptsize>1 on 64-bit machines   * undo commit 3963e46562e8   * set argument for ioctl FIONBIO to int   * fix for bug http://www.mail-archive.com/naviserver-devel@lists.sourceforge.net/msg02210.html

+++ nsd/sock.c
@@ -665,7 +665,7 @@ Ns_SockTimedConnect2(char *host, int port, char *lhost, int lport,
 int
 Ns_SockSetNonBlocking(SOCKET sock)
 {
-    unsigned long nb = 1;
+    int nb = 1;
 
     if (ns_sockioctl(sock, FIONBIO, &nb) == -1) {
         return NS_ERROR;
@@ -694,7 +694,7 @@ Ns_SockSetNonBlocking(SOCKET sock)
 int
 Ns_SockSetBlocking(SOCKET sock)
 {
-    unsigned long nb = 0;
+    int nb = 0;
 
     if (ns_sockioctl(sock, FIONBIO, &nb) == -1) {
         return NS_ERROR;
